add registration on event endpoints
register on event
cancel registration
check registration on event

+++ backend/presentations/app.py
@@ -1,12 +1,14 @@
 from typing import List
 from fastapi.responses import JSONResponse
 from fastapi_mail import FastMail, MessageSchema, ConnectionConfig, MessageType
-from fastapi import FastAPI, Path, status, HTTPException
+from fastapi import FastAPI, Path, status, HTTPException, Security
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.security import APIKeyHeader
 from repositories.db.event_repository import EventRepository
 from sqlalchemy.exc import OperationalError, ArgumentError
 from datetime import datetime
 from pydantic import BaseModel, EmailStr
+from services.JWTservice import TokenValidationResult, check_access_token
 
 
 app = FastAPI(
@@ -50,10 +52,17 @@ def allow_requests() -> None:
 
 
 @app.post("/")
-async def post_event(event: Event) -> None:
+async def post_event(event: Event, authorization_header: str = Security(APIKeyHeader(name='Authorization', auto_error=False))) -> None:
     """
     создание мероприятия
     """
+    token = check_access_token(authorization_header)
+    if token.exception is not None:
+        raise token.exception
+    
+    if not token.data['is_admin']:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="создавать мероприятия может только администратор")
+
 
     start_datetime = datetime(event.start_year, event.start_month, event.start_day, event.start_hour, event.start_minute)
     end_datetime = datetime(event.end_year, event.end_month, event.end_day, event.end_hour, event.end_minute)
@@ -82,10 +91,17 @@ async def get_event(id: str = Path(...)) -> dict | None:
 
 
 @app.put("/{id}")
-async def put_event(event: Event, id: str = Path(...)) -> None:
+async def put_event(event: Event, id: str = Path(...), authorization_header: str = Security(APIKeyHeader(name='Authorization', auto_error=False))) -> None:
     """
     обновление информации о мероприятии
     """
+    token = check_access_token(authorization_header)
+    if token.exception is not None:
+        raise token.exception
+    
+    if not token.data['is_admin']:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="обновлять мероприятия может только администратор")
+
 
     start_datetime = datetime(event.start_year, event.start_month, event.start_day, event.start_hour, event.start_minute)
     end_datetime = datetime(event.end_year, event.end_month, event.end_day, event.end_hour, event.end_minute)
@@ -99,10 +115,17 @@ async def put_event(event: Event, id: str = Path(...)) -> None:
 
 
 @app.delete("/{id}")
-async def delete_event(id: str = Path(...)) -> None:
+async def delete_event(id: str = Path(...), authorization_header: str = Security(APIKeyHeader(name='Authorization', auto_error=False))) -> None:
     """
     Удаление мероприятия
     """
+    token = check_access_token(authorization_header)
+    if token.exception is not None:
+        raise token.exception
+    
+    if not token.data['is_admin']:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="удалять мероприятия может только администратор")
+    
     try:
         await event_rep.delete_event(id)
     except OperationalError:
@@ -124,16 +147,32 @@ async def get_all_events() -> list:
 
 
 @app.post("/{id}")
-async def register_on_event(id: str = Path(...)) -> None:
+async def register_on_event(id: str = Path(...), authorization_header: str = Security(APIKeyHeader(name='Authorization', auto_error=False))) -> dict:
     """
     регистрация на мероприятие
     """
-    ...
+    token = check_access_token(authorization_header)
+    if token.exception is not None:
+        raise token.exception
+    
+    if not event_rep.check_registration(id, token.data['id']):
+        return {"detail": "already registered"}
+    
+    detail = await event_rep.register_on_event(id, token.data['id'])
+    return detail
 
 
 @app.delete("/{id}")
-async def cancel_registration(id: str = Path(...)) -> None:
+async def cancel_registration(id: str = Path(...), authorization_header: str = Security(APIKeyHeader(name='Authorization', auto_error=False))) -> dict:
     """
     отмена регистрации на мероприятие
     """
-    ...
+    token = check_access_token(authorization_header)
+    if token.exception is not None:
+        raise token.exception
+    
+    if event_rep.check_registration(id, token.data['id']):
+        detail = await event_rep.cancel_registration(id, token.data['id'])
+        return detail
+    else:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='пользователь не зарегистрирован на мероприятие')
